+++ lib/src/chart/line_chart/line_chart_painter.dart
@@ -33,8 +33,6 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       ..color = const Color(0x00000000)
       ..blendMode = BlendMode.dstIn;
 
-    _extraLinesPaint = Paint()..style = PaintingStyle.stroke;
-
     _touchLinePaint = Paint()
       ..style = PaintingStyle.stroke
       ..color = Colors.black;
@@ -42,8 +40,6 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
     _bgTouchTooltipPaint = Paint()
       ..style = PaintingStyle.fill
       ..color = Colors.white;
-
-    _imagePaint = Paint();
 
     _borderTouchTooltipPaint = Paint()
       ..style = PaintingStyle.stroke
@@ -54,10 +50,8 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
   late Paint _barAreaPaint;
   late Paint _barAreaLinesPaint;
   late Paint _clearBarAreaPaint;
-  late Paint _extraLinesPaint;
   late Paint _touchLinePaint;
   late Paint _bgTouchTooltipPaint;
-  late Paint _imagePaint;
   late Paint _borderTouchTooltipPaint;
 
   /// Paints [LineChartData] into the provided canvas.
@@ -93,7 +87,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
     }
 
     if (!data.extraLinesData.extraLinesOnTop) {
-      drawExtraLines(context, canvasWrapper, holder);
+      super.drawExtraLines(context, canvasWrapper, holder);
     }
 
     final lineIndexDrawingInfo = <LineIndexDrawingInfo>[];
@@ -110,7 +104,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       drawDots(canvasWrapper, barData, holder);
 
       if (data.extraLinesData.extraLinesOnTop) {
-        drawExtraLines(context, canvasWrapper, holder);
+        super.drawExtraLines(context, canvasWrapper, holder);
       }
 
       final indicatorsData = data.lineTouchData
@@ -959,159 +953,6 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
 
     barPath = barPath.toDashedPath(barData.dashArray);
     canvasWrapper.drawPath(barPath, _barPaint);
-  }
-
-  @visibleForTesting
-  void drawExtraLines(
-    BuildContext context,
-    CanvasWrapper canvasWrapper,
-    PaintHolder<LineChartData> holder,
-  ) {
-    final data = holder.data;
-    final viewSize = canvasWrapper.size;
-
-    if (data.extraLinesData.horizontalLines.isNotEmpty) {
-      for (final line in data.extraLinesData.horizontalLines) {
-        final from = Offset(0, getPixelY(line.y, viewSize, holder));
-        final to = Offset(viewSize.width, getPixelY(line.y, viewSize, holder));
-
-        _extraLinesPaint
-          ..color = line.color
-          ..strokeWidth = line.strokeWidth
-          ..transparentIfWidthIsZero();
-
-        canvasWrapper.drawDashedLine(
-          from,
-          to,
-          _extraLinesPaint,
-          line.dashArray,
-        );
-
-        if (line.sizedPicture != null) {
-          final centerX = line.sizedPicture!.width / 2;
-          final centerY = line.sizedPicture!.height / 2;
-          final xPosition = centerX;
-          final yPosition = to.dy - centerY;
-
-          canvasWrapper
-            ..save()
-            ..translate(xPosition, yPosition)
-            ..drawPicture(line.sizedPicture!.picture)
-            ..restore();
-        }
-
-        if (line.image != null) {
-          final centerX = line.image!.width / 2;
-          final centerY = line.image!.height / 2;
-          final centeredImageOffset = Offset(centerX, to.dy - centerY);
-          canvasWrapper.drawImage(
-            line.image!,
-            centeredImageOffset,
-            _imagePaint,
-          );
-        }
-
-        if (line.label.show) {
-          final label = line.label;
-          final style =
-              TextStyle(fontSize: 11, color: line.color).merge(label.style);
-          final padding = label.padding as EdgeInsets;
-
-          final span = TextSpan(
-            text: label.labelResolver(line),
-            style: Utils().getThemeAwareTextStyle(context, style),
-          );
-
-          final tp = TextPainter(
-            text: span,
-            textDirection: TextDirection.ltr,
-          )..layout();
-          canvasWrapper.drawText(
-            tp,
-            label.alignment.withinRect(
-              Rect.fromLTRB(
-                from.dx + padding.left,
-                from.dy - padding.bottom - tp.height,
-                to.dx - padding.right - tp.width,
-                to.dy + padding.top,
-              ),
-            ),
-          );
-        }
-      }
-    }
-
-    if (data.extraLinesData.verticalLines.isNotEmpty) {
-      for (final line in data.extraLinesData.verticalLines) {
-        final from = Offset(getPixelX(line.x, viewSize, holder), 0);
-        final to = Offset(getPixelX(line.x, viewSize, holder), viewSize.height);
-
-        _extraLinesPaint
-          ..color = line.color
-          ..strokeWidth = line.strokeWidth
-          ..transparentIfWidthIsZero();
-
-        canvasWrapper.drawDashedLine(
-          from,
-          to,
-          _extraLinesPaint,
-          line.dashArray,
-        );
-
-        if (line.sizedPicture != null) {
-          final centerX = line.sizedPicture!.width / 2;
-          final centerY = line.sizedPicture!.height / 2;
-          final xPosition = to.dx - centerX;
-          final yPosition = viewSize.height - centerY;
-
-          canvasWrapper
-            ..save()
-            ..translate(xPosition, yPosition)
-            ..drawPicture(line.sizedPicture!.picture)
-            ..restore();
-        }
-        if (line.image != null) {
-          final centerX = line.image!.width / 2;
-          final centerY = line.image!.height / 2;
-          final centeredImageOffset =
-              Offset(to.dx - centerX, viewSize.height - centerY);
-          canvasWrapper.drawImage(
-            line.image!,
-            centeredImageOffset,
-            _imagePaint,
-          );
-        }
-
-        if (line.label.show) {
-          final label = line.label;
-          final style =
-              TextStyle(fontSize: 11, color: line.color).merge(label.style);
-          final padding = label.padding as EdgeInsets;
-
-          final span = TextSpan(
-            text: label.labelResolver(line),
-            style: Utils().getThemeAwareTextStyle(context, style),
-          );
-
-          final tp = TextPainter(
-            text: span,
-            textDirection: TextDirection.ltr,
-          )..layout();
-
-          canvasWrapper.drawText(
-            tp,
-            label.alignment.withinRect(
-              Rect.fromLTRB(
-                to.dx - padding.right - tp.width,
-                from.dy + padding.top,
-                from.dx + padding.left,
-                to.dy - padding.bottom,
-              ),
-            ),
-          );
-        }
-      }
-    }
   }
 
   @visibleForTesting